+++ vedo/applications.py
@@ -879,6 +879,10 @@ class MorphByLandmarkPlotter(Plotter):
         if evt.picked3d is None:
             return
         self.current_pid = self.vtargets.closest_point(evt.picked3d, return_point_id=True)
+        if self.current_pid < 0:
+            self.status.text("No point exists. Press o to place a point.")
+            self.render()
+            return
         self.current_pt = self.targets[self.current_pid]
         self.status.text(
             f"Selected point ID {self.current_pid}" 
@@ -901,7 +905,10 @@ class MorphByLandmarkPlotter(Plotter):
         elif k == "o":
             if evt.object.name == "morphable_mesh":
                 self.sources.append(evt.picked3d)
-                self._sources = np.append(self._sources, [evt.picked3d], axis=0)
+                if len(self._sources) == 0:
+                    self._sources = np.array([evt.picked3d], dtype=float)
+                else:
+                    self._sources = np.append(self._sources, [evt.picked3d], axis=0)
                 self.targets.append(evt.picked3d)
                 self.status.text(
                     f"Added point {precision(evt.picked3d, 3)}. "
@@ -921,6 +928,42 @@ class MorphByLandmarkPlotter(Plotter):
 
                 self.remove("vsources", "vtargets", "vlines")
                 self.add(self.vsources, self.vtargets, self.vlines).render()
+
+        elif k == "c":  # clear all points
+            self.sources.clear()
+            self.targets.clear()
+            self._sources = np.array(self.sources, dtype=float)
+            self.vsources = Points(self._sources, r=self.point_size - 1)
+            self.vsources.c("red6", alpha=0.5).pickable(0)
+            self.vsources.name = "vsources"
+            self.vtargets = Points(self.targets, r=self.point_size)
+            self.vtargets.c("blue6", alpha=0.5).pickable(0)
+            self.vtargets.name = "vtargets"
+            self.vlines = vedo.Lines(self._sources, self.targets, c="k4").pickable(0)
+            self.vlines.name = "vlines"
+            self.remove("vsources", "vtargets", "vlines", "labels")
+            self.add(self.vsources, self.vtargets, self.vlines).render()
+
+        elif k == "d":  # delete the last point pair
+            if len(self.sources) == 0 or len(self.targets) == 0:
+                self.status.text("No points to delete.")
+                return
+            if len(self.sources) == len(self.targets) + 1:
+                self.sources.pop()
+            elif len(self.targets) == len(self.sources) + 1:
+                self.targets.pop()
+
+            self._sources = np.array(self.sources, dtype=float)
+            self.vsources = Points(self._sources, r=self.point_size - 1)
+            self.vsources.c("red6", alpha=0.5).pickable(0)
+            self.vsources.name = "vsources"
+            self.vtargets = Points(self.targets, r=self.point_size)
+            self.vtargets.c("blue6", alpha=0.5).pickable(0)
+            self.vtargets.name = "vtargets"
+            self.vlines = vedo.Lines(self._sources, self.targets, c="k4").pickable(0)
+            self.vlines.name = "vlines"
+            self.remove("vsources", "vtargets", "vlines", "labels")
+            self.add(self.vsources, self.vtargets, self.vlines).render()
 
         elif k in self.move_keys:  # use arrows to move the picked point
             if self.current_pid is None:
@@ -988,6 +1031,185 @@ class MorphByLandmarkPlotter(Plotter):
                 self.mesh.write(self.output_filename)
                 self.status.text(f"Morphed mesh saved to {self.output_filename}")
                 self.render()
+
+########################################################################################
+class MorphBySplinesPlotter(Plotter):
+
+    def __init__(self, meshes, **kwargs):
+        """
+        A Plotter to morph a mesh by moving points on it using splines.
+        It is useful in shape analysis, morphometrics, and mesh deformation.
+
+        Arguments:
+            meshes : (list of Mesh)
+                the mesh set to be matched
+            **kwargs : (dict)
+                keyword arguments to pass to a `vedo.plotter.Plotter` instance.
+        
+        Example:
+            ```python
+            from vedo import Mesh, dataurl
+            from vedo.applications import MorphBySplinesPlotter
+
+            m0 = Mesh(dataurl+"250.vtk").subdivide().c("o8", 0.5)
+            mm = Mesh(dataurl+"270.vtk").subdivide().c("k5", 0.5).x(800)
+            m1 = Mesh(dataurl+"290.vtk").subdivide().c("g8", 0.5).x(2000)
+
+            plt = MorphBySplinesPlotter([m0, mm, m1], size=(2050, 1250), axes=14)
+            plt.n_intermediates = 5  # number of intermediate shapes to generate
+            plt.show().close()
+            ```
+        """
+        vedo.settings.enable_default_keyboard_callbacks = False
+        vedo.settings.enable_default_mouse_callbacks = False
+        
+        super().__init__(**kwargs)
+
+        self.parallel_projection(True)
+        self.add_callback("key press", self._func)
+
+        self.n_intermediates = 10  # number of intermediate shapes to generate
+        self.mode = "2d"  # "2d", "3d" morphing mode
+
+        self.splinetools = []
+        self.intermediates = []
+        self.meshes = meshes
+        self.instructions = Text2D(
+            "     --- Morphing by Splines ---\n"
+            "Ctrl+f to fit the splines to the mesh.\n"
+            "Ctrl+a to add a new spline.\n"
+            "Ctrl+g to generate intermediate shapes.\n"
+            "Ctrl+s to save the splines.\n"
+            "Ctrl+l to load the splines from file.\n"
+            "Ctrl+z to remove the last spline.\n"
+            "Ctrl+c to clear all splines.\n"
+            "Ctrl+q to quit.",
+            pos="top-left",
+            bg="blue4",
+            alpha=0.1,
+            font="Calco",
+            s=0.75,
+        )
+        self.status = Text2D(
+            "Use Ctrl+a to add a new spline, or Ctrl+l to load splines from file.",
+            pos="bottom-left",
+            font="Calco",
+            bg="yellow5",
+            alpha=0.5,
+            s=1.1,
+        )
+        b = meshes[0].bounds()
+        self.offset = [0, (b[3] - b[2])/2, 0]
+        self.output_prefix = "morphed_"
+        self.output_suffix = ".vtk"
+        self.output_spline_filename = f"{self.output_prefix}splines.npy"
+        self.output_offset = 0  # used to generate an offset for the output filenames
+        self.add(meshes, self.status, self.instructions)
+
+
+    def _func(self, evt):
+
+        n = len(self.meshes)
+
+        if evt.keypress in ["q", "Ctrl+q"]:
+            self.break_interaction()
+
+        elif evt.keypress == "Ctrl+a":
+            if len(self.splinetools) > 0:
+                init_nodes = self.splinetools[-1].nodes() + self.offset
+            else:
+                b = self.meshes[0].bounds()
+                init_nodes = vedo.Line(
+                    [b[0], b[2], (b[4] + b[5]) / 2],
+                    [b[1], b[3], (b[4] + b[5]) / 2], res=n).coordinates
+            st = self.add_spline_tool(
+                init_nodes,
+                lc=len(self.splinetools),
+                lw=3,
+                can_add_nodes=False,
+            )
+            self.splinetools.append(st)
+            self.status.text(f"Added spline.")
+            self.render()
+
+        elif evt.keypress == "Ctrl+f":
+            for st in self.splinetools:
+                nodes = st.nodes()
+                for i in range(len(nodes)):
+                    nodes[i] = self.meshes[i].closest_point(nodes[i])
+                st.set_nodes(nodes)
+            self.status.text("Snapped spline points to meshes.")
+            self.render()
+
+        elif evt.keypress == "Ctrl+s":
+            all_nodes = []
+            for st in self.splinetools:
+                all_nodes.append(st.nodes())
+            all_nodes = np.array(all_nodes)
+            fname = f"{self.output_prefix}splines.npy"
+            np.save(fname, all_nodes)
+            for i, mi in enumerate(self.intermediates):
+                mi.write(f"{self.output_prefix}{self.output_offset+i}{self.output_suffix}")
+            self.status.text(
+                f"Saved {len(self.splinetools)} splines to {fname}.\n"
+                f"Saved {len(self.intermediates)} intermediate meshes saved as {self.output_prefix}*{self.output_suffix}")
+            self.render()
+
+        elif evt.keypress == "Ctrl+l":
+            if not os.path.exists(f"{self.output_prefix}splines.npy"):
+                self.status.text(f"File {self.output_prefix}splines.npy does not exist.")
+                self.render()
+                return
+            all_nodes = np.load(f"{self.output_prefix}splines.npy")
+            for st in self.splinetools:
+                st.off()
+            self.splinetools.clear()
+            for nodes in all_nodes:
+                st = self.add_spline_tool(
+                    nodes,
+                    lc=len(self.splinetools),
+                    lw=3,
+                    can_add_nodes=False,
+                )
+                self.splinetools.append(st)
+            self.status.text(
+                f"Loaded {len(self.splinetools)} splines from {self.output_prefix}splines.npy")
+            self.render()
+
+        elif evt.keypress == "Ctrl+z":
+            if len(self.splinetools) > 0:
+                self.splinetools[-1].off()
+                self.splinetools.pop()
+                self.status.text(f"Removed last spline.")
+                self.render()
+        
+        elif evt.keypress == "Ctrl+c":
+            self.remove(self.intermediates)
+            self.intermediates.clear()
+            self.status.text("Cleared all intermediate shapes.")
+            self.render()
+
+        elif evt.keypress == "Ctrl+g":
+            all_pts = []
+            for st in self.splinetools:
+                nodes = st.nodes()
+                coords = Spline(nodes, res=self.n_intermediates+1).coordinates
+                all_pts.append(coords)
+            all_pts = np.array(all_pts)
+            self.remove(self.intermediates)
+            self.intermediates.clear()
+            if len(all_pts) == 0:
+                self.status.text("No splines defined. Use Ctrl+a to add a spline.")
+                self.render()
+                return
+            warped0 = self.meshes[0].clone().wireframe().lighting("off").c("k4").alpha(1/n/5)
+            for i in range(1, self.n_intermediates+1):
+                sources = all_pts[:, i - 1]
+                targets = all_pts[:, i]
+                warped0.warp(sources, targets, mode=self.mode)
+                self.intermediates.append(warped0.clone())
+            self.status.text(f"Generated {len(self.intermediates)} intermediate shapes.")
+            self.add(self.intermediates).render()
 
 
 ########################################################################################